# SMM-Dashboard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SMM Panel API - Perfect Follow
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /platforms:
    get:
      operationId: getPlatforms
      tags: [platforms]
      summary: Get all platforms
      responses:
        "200":
          description: List of platforms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Platform"
    post:
      operationId: createPlatform
      tags: [platforms]
      summary: Create a platform (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlatformBody"
      responses:
        "201":
          description: Platform created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
  /platforms/{id}:
    get:
      operationId: getPlatform
      tags: [platforms]
      summary: Get a platform by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
    put:
      operationId: updatePlatform
      tags: [platforms]
      summary: Update a platform (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlatformBody"
      responses:
        "200":
          description: Platform updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
    delete:
      operationId: deletePlatform
      tags: [platforms]
      summary: Delete a platform (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Platform deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Platform:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        servicesCount:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - icon
        - servicesCount
        - status
        - color
        - createdAt
    CreatePlatformBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        status:
          type: string
          enum: [active, inactive]
        color:
          type: string
      required:
        - name
        - icon
        - status
        - color
...