# SMM-Dashboard/lib/api-zod/src/generated/types/createPlatformBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SMM Panel API - Perfect Follow
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePlatformBodyStatus } from "./createPlatformBodyStatus";

export interface CreatePlatformBody {
  name: string;
  icon: string;
  status: CreatePlatformBodyStatus;
  color: string;
}